requirements.txt
# Core ML & Data Processing
mlflow>=2.13.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
matplotlib>=3.7.0
openpyxl>=3.1.0

# Deep Learning (EfficientNet)
tensorflow>=2.15.0
keras>=2.15.0

# API & Serving
flask>=3.0.0
gunicorn>=21.2.0  # production WSGI server

# Image Processing
pillow>=10.0.0

# Dev & Testing
pytest>=8.0.0
ruff>=0.5.0
black>=24.0.0
